+++ core/core/Cargo.toml
@@ -102,7 +102,6 @@ services-gridfs = ["dep:mongodb", "dep:mongodb-internal-macros"]
 services-hdfs = ["dep:hdrs"]
 services-http = []
 services-huggingface = []
-services-ipfs = ["dep:prost"]
 services-ipmfs = []
 services-koofr = []
 services-lakefs = []
