+++ core/src/lib.rs
@@ -43,6 +43,8 @@ pub mod services {
     pub use opendal_service_ghac::*;
     #[cfg(feature = "services-hdfs-native")]
     pub use opendal_service_hdfs_native::*;
+    #[cfg(feature = "services-ipfs")]
+    pub use opendal_service_ipfs::*;
     #[cfg(feature = "services-moka")]
     pub use opendal_service_moka::*;
     #[cfg(feature = "services-mysql")]
